// hub/frontend/lib/api/generated/model/auditEventRead.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Wardn Hub API
 * OpenAPI spec version: 0.2.122
 */
import type { AuditEventReadMetadata } from './auditEventReadMetadata';

export interface AuditEventRead {
  actorTokenId?: string | null;
  actorUserId?: string | null;
  createdAt: string;
  eventType: string;
  id: string;
  ipAddress: string;
  metadata?: AuditEventReadMetadata;
  organizationId?: string | null;
  subjectId: string;
  subjectType: string;
  userAgent: string;
}