// hub/frontend/lib/api/generated/model/userRead.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Wardn Hub API
 * OpenAPI spec version: 0.2.90
 */

export interface UserRead {
  created_at: string;
  display_name: string;
  email: string;
  first_name: string;
  id: string;
  is_active: boolean;
  is_global_moderator: boolean;
  is_global_partner_manager: boolean;
  is_superuser: boolean;
  last_login_at: string | null;
  last_name: string;
  updated_at: string;
}